requirements.txt
alembic==1.5.8
aniso8601==9.0.1
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.15.6
PyMySQL==1.0.2
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.3
Werkzeug==0.16.1
